tpm12: Cast sizeof result to uint32_t to match va_arg type (-fanalyzer)
Resolve the following gcc -fanalyzer issue by casting the result of
sizeof() to uint32_t.

tpm12/tpm_cryptoh.c:977:16: warning: 'va_arg' expected 'uint32_t' \
   {aka 'unsigned int'} but received 'long unsigned int' \
   for variadic argument 3 of 'ap' [CWE-686] [-Wanalyzer-va-arg-type-mismatch]

+++ src/tpm12/tpm_cryptoh.c
@@ -2374,7 +2374,7 @@ TPM_RESULT TPM_CryptoTest(void)
     if (rc == 0) {
 	printf(" TPM_CryptoTest: Test 1 - SHA1 one part\n");
 	rc = TPM_SHA1(actual,
-		      sizeof(buffer1) - 1, buffer1,
+		      (uint32_t)sizeof(buffer1) - 1, buffer1,
 		      0, NULL);
     }
     if (rc == 0) {
@@ -2390,7 +2390,7 @@ TPM_RESULT TPM_CryptoTest(void)
 	printf(" TPM_CryptoTest: Test 2 - SHA1 two parts\n");
 	rc = TPM_SHA1(actual,
 		      16, buffer1,	/* first 16 */
-		      sizeof(buffer1) - 17, buffer1 + 16,	/* rest */
+		      (uint32_t)sizeof(buffer1) - 17, buffer1 + 16,	/* rest */
 		      0, NULL);
     }
     if (rc == 0) {
